requirements.txt
pillow
moviepy
pyautogui
pyperclip
pyttsx3
requests
